refactor to extract cosmic text logic
